requirements.txt
requests
protobuf
pycryptodome
zstandard
pyyaml